fixed new navbar login/logout btn
